+++ CHANGELOG.rst
@@ -11,6 +11,9 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- Added ``put_metadata_schema_item`` method to
+  ``dtool_gui_tk.models.MetadataSchemaListModel``
+
 
 Changed
 ^^^^^^^